refactor(lowlevel): move echo test API to httpapi package
